chore: split cache.stats and extract warm manifests to lift cache.py off the allowlist
Two changes that together take src/mgz_pkmn/cache.py off both RADON_CC_EXCLUDE
and RADON_MI_EXCLUDE — emptying the radon allowlist entirely (epic #551 reaches
zero).

CC: stats() was D-21. Split each cache slice into its own helper —
_scan_json_dir (hoisted from a nested def), _override_stats, and one
_*_warm_stats per warm manifest, backed by small _warm_timestamp / _warm_int /
_warm_bool field readers. stats() is now an A-1 coordinator that stitches the
pieces into a CacheStats.

MI: the file's maintainability index was B (14.23), dominated by its ~1.6K LOC.
Moved the five warm-manifest slices (concept / set-cards / sets / cards / images
— their constants plus the _path/read/write/is_fresh quartets, ~420 lines) into
a new _cache_warm.py submodule, re-exported from cache.py so the public surface
(cache.read_concept_warm, cache.SETS_WARM_SCHEMA_VERSION, …) is unchanged. The
back-import sits at the bottom of each module so the cycle resolves regardless
of which side is imported first; the slices touch cache's core path helpers
(cache_root / _cache_root_path) only at call time. cache.py is now A (24.96),
_cache_warm.py A (49.53).

Empties RADON_CC_EXCLUDE and RADON_MI_EXCLUDE in the Makefile.
